refactor(services): extract instance_validation module
Track D.11.  Lift the pure validation helpers and their form-output
dataclasses (ConnectionCheck, SearchModes, validate_cutoff_controls,
validate_upgrade_controls, resolve_search_modes, type_mismatch_message,
and the private _whisparr_version_major helper plus the
_APP_NAME_TO_TYPE map) out of routes/settings/_helpers.py into a
new houndarr.services.instance_validation module.

Why split now: the helpers are pure logic.  Living under routes/
meant they imported through the route layer, which cost clarity
during the D.10 submit-service extraction.  With them in the
service layer the submit service depends on a sibling service,
not back across a layer boundary.  check_connection stays in
_helpers because it performs a live HTTP probe through the client
layer and is not pure; the returned ConnectionCheck type is the
service's one, so callers only see a single dataclass.

_helpers.py re-exports the moved names via __all__ so existing
imports through that module keep working; instance_submit.py
updates to import the validators from the service directly.

just quick green (1986 passed); mypy strict clean.
