+++ src/houndarr/routes/settings/_helpers.py
@@ -12,7 +12,6 @@ from __future__ import annotations
 
 import html
 import logging
-from dataclasses import dataclass
 
 from fastapi import Request
 from fastapi.responses import HTMLResponse
@@ -46,6 +45,14 @@ from houndarr.config import (
 )
 from houndarr.routes._htmx import is_hx_request
 from houndarr.routes._templates import get_templates
+from houndarr.services.instance_validation import (
+    ConnectionCheck,
+    SearchModes,
+    resolve_search_modes,
+    type_mismatch_message,
+    validate_cutoff_controls,
+    validate_upgrade_controls,
+)
 from houndarr.services.instances import (
     Instance,
     InstanceType,
@@ -57,7 +64,6 @@ from houndarr.services.instances import (
     list_instances,
 )
 
-
 # active_error_instance_ids() is added by PR18 (f65e0bc).  Until it lands,
 # return an empty set so template `(active_error_ids or [])` checks
 # evaluate to "no instances in error" without breaking the call sites.
@@ -65,6 +71,29 @@ async def active_error_instance_ids() -> set[int]:
     """Stub: returns empty until PR18 ships the real query."""
     return set()
 
+
+# Re-exported from houndarr.services.instance_validation so existing
+# imports from this module keep working; D.11 moved the real definitions
+# into the service layer.
+__all__ = [
+    "API_KEY_UNCHANGED",
+    "ArrClient",
+    "ConnectionCheck",
+    "SearchModes",
+    "active_error_instance_ids",
+    "blank_instance",
+    "build_client",
+    "check_connection",
+    "connection_guard_response",
+    "connection_status_response",
+    "master_key",
+    "render",
+    "render_settings_page",
+    "resolve_search_modes",
+    "type_mismatch_message",
+    "validate_cutoff_controls",
+    "validate_upgrade_controls",
+]
 
 logger = logging.getLogger(__name__)
 
@@ -145,32 +174,19 @@ def build_client(instance_type: InstanceType, url: str, api_key: str) -> ArrClie
     return client_cls(url=url, api_key=api_key)
 
 
-@dataclass(frozen=True, slots=True)
-class ConnectionCheck:
-    """Result of a connection test against an *arr instance."""
-
-    reachable: bool
-    app_name: str | None = None
-    version: str | None = None
-
-
-_APP_NAME_TO_TYPE: dict[str, InstanceType] = {
-    "radarr": InstanceType.radarr,
-    "sonarr": InstanceType.sonarr,
-    "lidarr": InstanceType.lidarr,
-    "readarr": InstanceType.readarr,
-    # Whisparr v2 and v3 both report appName "Whisparr"; version-based
-    # disambiguation is handled in type_mismatch_message.
-    "whisparr": InstanceType.whisparr_v2,
-}
-
-
 async def check_connection(
     instance_type: InstanceType,
     url: str,
     api_key: str,
 ) -> ConnectionCheck:
-    """Test connectivity and identify the remote *arr application."""
+    """Test connectivity and identify the remote *arr application.
+
+    Kept here (and not in :mod:`houndarr.services.instance_validation`)
+    because it performs a live HTTP probe through the client layer;
+    the validation service stays pure.  The returned
+    :class:`ConnectionCheck` is the service type so callers only see
+    one dataclass for the connection-probe result.
+    """
     client = build_client(instance_type, url, api_key)
     async with client:
         status = await client.ping()
@@ -181,49 +197,6 @@ async def check_connection(
         app_name=status.app_name,
         version=status.version,
     )
-
-
-def _whisparr_version_major(version: str | None) -> int | None:
-    """Extract the major version number from a Whisparr version string."""
-    if not version:
-        return None
-    try:
-        return int(version.split(".")[0])
-    except (ValueError, IndexError):
-        return None
-
-
-def type_mismatch_message(check: ConnectionCheck, selected: InstanceType) -> str | None:
-    """Return a mismatch error string, or ``None`` if the type is valid."""
-    if check.app_name is None:
-        return None
-
-    app_lower = check.app_name.lower()
-    detected = _APP_NAME_TO_TYPE.get(app_lower)
-
-    # Whisparr v2 and v3 both report appName "Whisparr". Use version to
-    # detect v3 (major version >= 3) and check against the selected type.
-    if app_lower == "whisparr":
-        major = _whisparr_version_major(check.version)
-        if major is not None and major >= 3 and selected == InstanceType.whisparr_v2:
-            return (
-                f"Version mismatch: this URL runs Whisparr v3 ({check.version})."
-                " Select 'Whisparr v3' as the instance type."
-            )
-        if major is not None and major < 3 and selected == InstanceType.whisparr_v3:
-            return (
-                f"Version mismatch: this URL runs Whisparr v2 ({check.version})."
-                " Select 'Whisparr v2' as the instance type."
-            )
-        # Correct pairing; skip the generic app-name check.
-        return None
-
-    if detected is None:
-        # Unknown app name (e.g. a Readarr fork); allow through.
-        return None
-    if detected != selected:
-        return f"Type mismatch: this URL is running {check.app_name}, not {selected.value.title()}."
-    return None
 
 
 def connection_status_response(message: str, ok: bool, status_code: int) -> HTMLResponse:
@@ -247,111 +220,6 @@ def connection_guard_response(message: str) -> HTMLResponse:
             "HX-Reswap": "innerHTML",
             "HX-Trigger": "houndarr-connection-test-failure",
         },
-    )
-
-
-def validate_cutoff_controls(
-    cutoff_batch_size: int,
-    cutoff_cooldown_days: int,
-    cutoff_hourly_cap: int,
-) -> str | None:
-    """Validate cutoff-specific numeric controls from form submissions."""
-    if cutoff_batch_size < 1:
-        return "Cutoff batch size must be at least 1."
-    if cutoff_cooldown_days < 0:
-        return "Cutoff cooldown days must be 0 or greater."
-    if cutoff_hourly_cap < 0:
-        return "Cutoff hourly cap must be 0 or greater."
-    return None
-
-
-def validate_upgrade_controls(
-    upgrade_batch_size: int,
-    upgrade_cooldown_days: int,
-    upgrade_hourly_cap: int,
-) -> str | None:
-    """Validate upgrade-specific numeric controls from form submissions."""
-    if upgrade_batch_size < 1:
-        return "Upgrade batch size must be at least 1."
-    if upgrade_cooldown_days < 7:
-        return "Upgrade cooldown days must be at least 7."
-    if upgrade_hourly_cap < 0:
-        return "Upgrade hourly cap must be 0 or greater."
-    return None
-
-
-class SearchModes:
-    """Resolved per-app search mode enum values."""
-
-    __slots__ = ("lidarr", "readarr", "sonarr", "whisparr")
-
-    def __init__(
-        self,
-        sonarr: SonarrSearchMode,
-        lidarr: LidarrSearchMode,
-        readarr: ReadarrSearchMode,
-        whisparr: WhisparrSearchMode,
-    ) -> None:
-        self.sonarr = sonarr
-        self.lidarr = lidarr
-        self.readarr = readarr
-        self.whisparr = whisparr
-
-
-def resolve_search_modes(
-    instance_type: InstanceType,
-    sonarr_raw: str,
-    lidarr_raw: str,
-    readarr_raw: str,
-    whisparr_raw: str,
-) -> SearchModes | str:
-    """Validate and resolve per-app search mode strings into enum values.
-
-    Returns a :class:`SearchModes` with validated values, or a plain error
-    string if any value is invalid.  Non-applicable search modes default to
-    their enum's first value.
-    """
-    try:
-        sonarr_mode = (
-            SonarrSearchMode(sonarr_raw)
-            if instance_type == InstanceType.sonarr
-            else SonarrSearchMode.episode
-        )
-    except ValueError:
-        return "Invalid Sonarr search mode."
-
-    try:
-        lidarr_mode = (
-            LidarrSearchMode(lidarr_raw)
-            if instance_type == InstanceType.lidarr
-            else LidarrSearchMode.album
-        )
-    except ValueError:
-        return "Invalid Lidarr search mode."
-
-    try:
-        readarr_mode = (
-            ReadarrSearchMode(readarr_raw)
-            if instance_type == InstanceType.readarr
-            else ReadarrSearchMode.book
-        )
-    except ValueError:
-        return "Invalid Readarr search mode."
-
-    try:
-        whisparr_mode = (
-            WhisparrSearchMode(whisparr_raw)
-            if instance_type == InstanceType.whisparr_v2
-            else WhisparrSearchMode.episode
-        )
-    except ValueError:
-        return "Invalid Whisparr search mode."
-
-    return SearchModes(
-        sonarr=sonarr_mode,
-        lidarr=lidarr_mode,
-        readarr=readarr_mode,
-        whisparr=whisparr_mode,
     )
